requirements.txt
streamlit
nltk
scikit-learn
sentence-transformers
torch
PyPDF2
streamlit-lottie
requests
Pillow
